Extract sortable bullet list
