+++ src/pages/content/components/ExperienceCard.tsx
@@ -1,25 +1,8 @@
-import { Bars2Icon } from "@heroicons/react/24/outline";
 import Checkbox from "../../../common/components/Checkbox";
 import EllipsisMenu from "../../../common/components/EllipsisMenu";
 import { Experience } from "../../../common/interfaces/resume";
-import {
-  closestCenter,
-  DndContext,
-  KeyboardSensor,
-  PointerSensor,
-  useSensor,
-  useSensors,
-} from "@dnd-kit/core";
-import {
-  arrayMove,
-  SortableContext,
-  sortableKeyboardCoordinates,
-  verticalListSortingStrategy,
-} from "@dnd-kit/sortable";
-import SortableItem from "./SortableItem";
-import { restrictToVerticalAxis } from "@dnd-kit/modifiers";
-import classNames from "classnames";
 import { monthYearToString } from "../../../common/functions/time";
+import SortableBulletList from "./SortableBulletList";
 
 interface ExperienceCardProps {
   title: string;
@@ -38,15 +21,6 @@ const ExperienceCard: React.FC<ExperienceCardProps> = ({
   onEdit,
   onDelete,
 }) => {
-  const sensors = useSensors(
-    useSensor(PointerSensor),
-    useSensor(KeyboardSensor, {
-      coordinateGetter: sortableKeyboardCoordinates,
-    })
-  );
-
-  const bulletIds = experience.bulletPoints.map((b) => b.id);
-
   const start = monthYearToString(experience.startDate);
   const end = experience.endDate
     ? monthYearToString(experience.endDate)
@@ -85,64 +59,10 @@ const ExperienceCard: React.FC<ExperienceCardProps> = ({
           />
         </div>
 
-        <DndContext
-          sensors={sensors}
-          collisionDetection={closestCenter}
-          onDragEnd={(e) => {
-            const { active, over } = e;
-
-            if (active.id !== over?.id) {
-              const oldIndex = bulletIds.indexOf(active.id as string);
-              const newIndex = bulletIds.indexOf(over?.id as string);
-
-              const newBullets = arrayMove(
-                experience.bulletPoints,
-                oldIndex,
-                newIndex
-              );
-
-              onChange({ ...experience, bulletPoints: newBullets });
-            }
-          }}
-          modifiers={[restrictToVerticalAxis]}
-        >
-          <SortableContext
-            items={bulletIds}
-            strategy={verticalListSortingStrategy}
-          >
-            <ul className="flex flex-col py-2">
-              {experience.bulletPoints.map((bulletPoint, index) => (
-                <SortableItem key={bulletPoint.id} id={bulletPoint.id}>
-                  <li className="py-2 px-4 flex gap-4 items-center">
-                    <Checkbox
-                      checked={bulletPoint.included}
-                      onChange={(e) => {
-                        const newBullets = experience.bulletPoints.slice();
-                        newBullets[index] = {
-                          ...newBullets[index],
-                          included: e.target.checked,
-                        };
-                        onChange({ ...experience, bulletPoints: newBullets });
-                      }}
-                    />
-                    <span
-                      className={classNames("text-sm", {
-                        "text-gray-700": bulletPoint.included,
-                        "text-gray-400": !bulletPoint.included,
-                      })}
-                    >
-                      {bulletPoint.text}
-                    </span>
-                    <Bars2Icon
-                      className="text-gray-400 flex-shrink-0 ml-auto h-4 w-8"
-                      aria-hidden="true"
-                    />
-                  </li>
-                </SortableItem>
-              ))}
-            </ul>
-          </SortableContext>
-        </DndContext>
+        <SortableBulletList
+          bullets={experience.bulletPoints}
+          onChange={(bulletPoints) => onChange({ ...experience, bulletPoints })}
+        />
       </div>
     </>
   );
